CultOfQoL: Fix god tear bug, float costs, logging standardization
Fix god tear double-collection by moving from GiveGodTearIE postfix
to OnInteract prefix (snapshots AbilityPoints before deferred callbacks).
Change mass action gold/time costs from int to float with 0.25 stepping.
Standardize all 21 untagged log calls across 7 files to consistent
[Tag] prefix format with detailed cost/faith/god tear logging.

+++ Cult of the Lamb/CultOfQoL/Configs.cs
@@ -149,8 +149,8 @@ public partial class Plugin
 
     internal static ConfigEntry<bool> MassBribe { get; private set; }
 
-    internal static ConfigEntry<int> MassActionGoldCost { get; private set; }
-    internal static ConfigEntry<int> MassActionTimeCost { get; private set; }
+    internal static ConfigEntry<float> MassActionGoldCost { get; private set; }
+    internal static ConfigEntry<float> MassActionTimeCost { get; private set; }
     internal static ConfigEntry<int> MassFaithReduction { get; private set; }
 
     internal static ConfigEntry<bool> RotFertilizerDecay { get; private set; }
